% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist_functions.R
\name{csp}
\alias{csp}
\title{Calculates the Colony Site Persistence Index}
\usage{
csp(xy1, xy2, area)
}
\arguments{
\item{xy1}{A set of coordinates at time 1}

\item{xy2}{A set of coordinates at time 2}

\item{area}{An area in the same units as xy}
}
\value{
The Colony Site Persistence Index
}
\description{
Function csp accepts a pair of xy-coordinates, representing colony
locations, at two different times (xy1 and xy2) and the area of the in which
those colony locations are distributed.  The arguments xy1 and xy2 should be
a matrix or a data-frame R objects consist of x and y values.  The unit of
the area should be the same as that for x and y values.  Function csp
then returns the colony site persistence index of two sets of colonies for two
consecutive points in time based on Carrasco and Toquenaga (in preparation).
The csp index averages the Standardized Mean Nearest Neighbour Distance
(smnnd) between the smnnd calculated from the xy1 colonies to the xy2
colonies, and the smnnd calculated from the xy2 colonies to the xy1 colonies.
A negative sign is added so that increasing values of cspindex represent
increases in colony site persistence.
}
\examples{
colonies1 <- matrix(sample(1:10,10),nrow=5,ncol=2)
colonies2 <- matrix(sample(1:10,10),nrow=5,ncol=2)
csp(colonies1,colonies2,100)
}
