% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist_functions.R
\name{mnnd}
\alias{mnnd}
\title{Calculates the Mean Nearest Neighbor Distance}
\usage{
mnnd(xy1, xy2 = NULL)
}
\arguments{
\item{xy1}{A set of coordinates}

\item{xy2}{A set of coordinates}
}
\value{
The Mean Nearest Neighbor Distance
}
\description{
Function mnnd accepts a xy-coordinate (xy1) or a pair of
xy-coordinates (xy1 and xy2). The arguments xy1 and xy2
should be a matrix or a data-frame R objects consist of x and y values.
Function mnnd then returns the mean nearest neighbor distance based on
Clark and Evans (1954) Ecology: 35(4): 445-453.
}
\examples{
coords1 <- matrix(sample(1:10,10),nrow=5,ncol=2)
cooords2 <- matrix(sample(1:10,10),nrow=5,ncol=2)
mnnd(coords1,cooords2)
}
