% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist_package.R
\docType{package}
\name{persist}
\alias{persist}
\alias{persist-package}
\title{persist: A package for estimating Colony Site Persistence.}
\description{
The persist package provides two functions for calculating nearest
neighbor distances, and one main function for calculating the Colony Site
Persistence Index based on Carrasco ant Toquenaga (in prep.)
}
