% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist_functions.R
\name{smnnd}
\alias{smnnd}
\title{Calculates the Mean Nearest Neighbor Distances}
\usage{
smnnd(xy1, xy2 = NULL, area)
}
\arguments{
\item{xy1}{A set of coordinates}

\item{xy2}{A set of coordinates}

\item{area}{An area in the same units as xy}
}
\value{
The Standardized Mean Nearest Neighbor Distance
}
\description{
Function smnnd accepts a xy-coordinate (xy1) or a pair of xy-coordinates
(xy1 and xy2) and the area of the arena in which those spatial points are
distributed.  The arguments xy1 and xy2 should be a matrix or a data-frame R
objects consist of x and y values. The unit of the area should be the same as
that for x and y values.
Function smnnd then returns the mean nearest neighbor distance
standardized with the Poisson expectation based on Clark and Evans (1954)
Ecology: 35(4): 445-453: E(nnd) = 1/sqrt(n)/2, where n is the number of
points within a square area with the sides range [0,1].  This expectation,
however, underestimates the distances when the number of points are very low
(<10), so a tentative correction term: 1/n/4, was added.  Standardization is
achieved by subtracting the corrected E(nnd) from the observed mean nearest
neighbor distance rescaled with sqrt(area).
}
\examples{
coords1 <- matrix(sample(1:10,10),nrow=5,ncol=2)
coords2 <- matrix(sample(1:10,10),nrow=5,ncol=2)
smnnd(coords1,coords2,100)
smnnd(coords2,coords1,100)
}
